requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.27
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.3
